refactor: relocate integration testing utils
